refactor(manufacture): extract FlexiManufactureMovementService
