+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/FlexiManufactureClient.cs
@@ -1,29 +1,15 @@
 using Anela.Heblo.Adapters.Flexi.Manufacture.Internal;
-using Anela.Heblo.Adapters.Flexi.Stock;
 using Anela.Heblo.Domain.Features.Catalog;
 using Anela.Heblo.Domain.Features.Catalog.Lots;
-using Anela.Heblo.Domain.Features.Catalog.Stock;
 using Anela.Heblo.Domain.Features.Manufacture;
 using Microsoft.Extensions.Logging;
-using Rem.FlexiBeeSDK.Client;
 using Rem.FlexiBeeSDK.Client.Clients.Accounting.Ledger;
 using Rem.FlexiBeeSDK.Client.Clients.Products.BoM;
-using Rem.FlexiBeeSDK.Client.Clients.Products.StockMovement;
-using Rem.FlexiBeeSDK.Model;
-using Rem.FlexiBeeSDK.Model.Products.StockMovement;
 
 namespace Anela.Heblo.Adapters.Flexi.Manufacture;
 
 internal sealed class FlexiManufactureClient : IManufactureClient
 {
-    private const string WarehouseDocumentType_OutboundMaterial = "V-VYDEJ-MATERIAL";
-    private const string WarehouseDocumentType_InboundSemiProduct = "V-PRIJEM-POLOTOVAR";
-    private const string WarehouseDocumentType_OutboundSemiProduct = "V-VYDEJ-POLOTOVAR";
-    private const string WarehouseDocumentType_InboundProduct = "V-PRIJEM-VYROBEK";
-
-
-    private readonly IErpStockClient _stockClient;
-    private readonly IStockItemsMovementClient _stockMovementClient;
     private readonly IBoMClient _bomClient;
     private readonly IProductSetsClient _productSetsClient;
     private readonly ILogger<FlexiManufactureClient> _logger;
@@ -32,10 +18,9 @@ internal sealed class FlexiManufactureClient : IManufactureClient
     private readonly IFlexiIngredientRequirementAggregator _requirementAggregator;
     private readonly IFlexiIngredientStockValidator _stockValidator;
     private readonly IFlexiLotLoader _lotLoader;
+    private readonly IFlexiManufactureMovementService _movementService;
 
     public FlexiManufactureClient(
-        IErpStockClient stockClient,
-        IStockItemsMovementClient stockMovementClient,
         IBoMClient bomClient,
         IProductSetsClient productSetsClient,
         ILogger<FlexiManufactureClient> logger,
@@ -43,10 +28,9 @@ internal sealed class FlexiManufactureClient : IManufactureClient
         IFefoConsumptionAllocator fefoAllocator,
         IFlexiIngredientRequirementAggregator requirementAggregator,
         IFlexiIngredientStockValidator stockValidator,
-        IFlexiLotLoader lotLoader)
+        IFlexiLotLoader lotLoader,
+        IFlexiManufactureMovementService movementService)
     {
-        _stockClient = stockClient ?? throw new ArgumentNullException(nameof(stockClient));
-        _stockMovementClient = stockMovementClient ?? throw new ArgumentNullException(nameof(stockMovementClient));
         _bomClient = bomClient;
         _productSetsClient = productSetsClient;
         _logger = logger ?? throw new ArgumentNullException(nameof(logger));
@@ -55,6 +39,7 @@ internal sealed class FlexiManufactureClient : IManufactureClient
         _requirementAggregator = requirementAggregator ?? throw new ArgumentNullException(nameof(requirementAggregator));
         _stockValidator = stockValidator ?? throw new ArgumentNullException(nameof(stockValidator));
         _lotLoader = lotLoader ?? throw new ArgumentNullException(nameof(lotLoader));
+        _movementService = movementService ?? throw new ArgumentNullException(nameof(movementService));
     }
 
     public async Task<string> SubmitManufactureAsync(SubmitManufactureClientRequest request, CancellationToken cancellationToken = default)
@@ -130,131 +115,10 @@ internal sealed class FlexiManufactureClient : IManufactureClient
         }
 
         // Phase 2: Create ONE consume document (per warehouse) with all consumption lines
-        await SubmitConsolidatedConsumptionMovementsAsync(request, allConsumptionItems, productCosts, cancellationToken);
+        await _movementService.SubmitConsolidatedConsumptionAsync(request, allConsumptionItems, productCosts, cancellationToken);
 
         // Phase 3: Create ONE produce document with all products
-        await SubmitConsolidatedProductionMovementAsync(request, productCosts, cancellationToken);
-    }
-
-    private async Task SubmitConsolidatedConsumptionMovementsAsync(
-        SubmitManufactureClientRequest request,
-        List<ConsumptionItem> consumptionItems,
-        Dictionary<string, double> productCosts,
-        CancellationToken cancellationToken)
-    {
-        var consumptionByWarehouse = consumptionItems
-            .GroupBy(item => FlexiWarehouseResolver.ForProductType(item.ProductType))
-            .ToList();
-
-        foreach (var warehouseGroup in consumptionByWarehouse)
-        {
-            var warehouseId = warehouseGroup.Key;
-            var stockItems = await _stockClient.StockToDateAsync(request.Date, warehouseId, cancellationToken);
-            var stockMovementItems = new List<StockItemsMovementUpsertRequestItemFlexiDto>();
-
-            foreach (var consumptionItem in warehouseGroup)
-            {
-                var stockItem = stockItems.FirstOrDefault(s => s.ProductCode == consumptionItem.ProductCode);
-                var unitPrice = stockItem != null ? (double)stockItem.Price : 0;
-
-                // Track cost per manufactured product
-                productCosts[consumptionItem.SourceProductCode] += unitPrice * consumptionItem.Amount;
-
-                stockMovementItems.Add(new StockItemsMovementUpsertRequestItemFlexiDto
-                {
-                    ProductCode = consumptionItem.ProductCode,
-                    ProductName = consumptionItem.ProductName,
-                    Amount = consumptionItem.Amount,
-                    AmountIssued = consumptionItem.Amount,
-                    LotNumber = consumptionItem.LotNumber,
-                    Expiration = consumptionItem.Expiration?.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc),
-                    UnitPrice = unitPrice,
-                });
-            }
-
-            var documentType = GetConsumptionDocumentType(warehouseId);
-            var note = CreateDescription(request);
-
-            var consumptionRequest = new StockItemsMovementUpsertRequestFlexiDto
-            {
-                CreatedBy = request.CreatedBy,
-                AccountingDate = request.Date,
-                IssueDate = request.Date,
-                StockItems = stockMovementItems,
-                Description = note,
-                DocumentTypeCode = documentType,
-                StockMovementDirection = StockMovementDirection.Out,
-                Note = request.ManufactureOrderCode,
-                WarehouseId = warehouseId.ToString()
-            };
-
-            var consumptionResult = await _stockMovementClient.SaveAsync(consumptionRequest, cancellationToken);
-
-            if (consumptionResult != null && !consumptionResult.IsSuccess)
-            {
-                throw new FlexiManufactureException(
-                    FlexiManufactureOperationKind.ConsumptionMovement,
-                    $"Failed to create consumption stock movement for warehouse {warehouseId}",
-                    warehouseId: warehouseId,
-                    rawFlexiError: consumptionResult.GetErrorMessage());
-            }
-        }
-    }
-
-    private async Task SubmitConsolidatedProductionMovementAsync(
-        SubmitManufactureClientRequest request,
-        Dictionary<string, double> productCosts,
-        CancellationToken cancellationToken)
-    {
-        var productMovementItems = request.Items
-            .Where(i => i.Amount > 0)
-            .Select(item =>
-            {
-                var productCost = productCosts.GetValueOrDefault(item.ProductCode, 0);
-                var unitPrice = item.Amount > 0 ? productCost / (double)item.Amount : 0;
-
-                return new StockItemsMovementUpsertRequestItemFlexiDto
-                {
-                    ProductCode = item.ProductCode,
-                    ProductName = item.ProductName,
-                    Amount = (double)item.Amount,
-                    AmountIssued = (double)item.Amount,
-                    LotNumber = request.LotNumber,
-                    Expiration = request.ExpirationDate?.ToDateTime(TimeOnly.MinValue, DateTimeKind.Utc),
-                    UnitPrice = unitPrice
-                };
-            }).ToList();
-
-        // Don't create production movement if there are no products
-        if (productMovementItems.Count == 0)
-        {
-            return;
-        }
-
-        var documentType = GetProductionDocumentType(request.ManufactureType);
-        var note = CreateDescription(request);
-        var productionRequest = new StockItemsMovementUpsertRequestFlexiDto
-        {
-            CreatedBy = request.CreatedBy,
-            AccountingDate = request.Date,
-            IssueDate = request.Date,
-            StockItems = productMovementItems,
-            Description = note,
-            DocumentTypeCode = documentType,
-            StockMovementDirection = StockMovementDirection.In,
-            Note = request.ManufactureOrderCode,
-            WarehouseId = FlexiStockClient.ProductsWarehouseId.ToString()
-        };
-
-        var productionResult = await _stockMovementClient.SaveAsync(productionRequest, cancellationToken);
-
-        if (productionResult != null && !productionResult.IsSuccess)
-        {
-            throw new FlexiManufactureException(
-                FlexiManufactureOperationKind.ProductionMovement,
-                "Failed to create production stock movement",
-                rawFlexiError: productionResult.GetErrorMessage());
-        }
+        await _movementService.SubmitConsolidatedProductionAsync(request, productCosts, cancellationToken);
     }
 
     public async Task UpdateBoMIngredientAmountAsync(string productCode, string ingredientCode, double newAmount, CancellationToken cancellationToken = default)
@@ -294,36 +158,6 @@ internal sealed class FlexiManufactureClient : IManufactureClient
                 Amount = s.Quantity,
             })
             .ToList();
-    }
-
-
-    private static string GetProductionDocumentType(ErpManufactureType manufactureType)
-    {
-        var documentType = manufactureType switch
-        {
-            ErpManufactureType.SemiProduct => WarehouseDocumentType_InboundSemiProduct,
-            ErpManufactureType.Product => WarehouseDocumentType_InboundProduct,
-            _ => throw new InvalidOperationException("Unknown warehouse for consumption movement for manufacture type " + manufactureType)
-        };
-        return documentType;
-    }
-
-    private static string GetConsumptionDocumentType(int warehouseId)
-    {
-        var documentType = warehouseId switch
-        {
-            FlexiStockClient.SemiProductsWarehouseId => WarehouseDocumentType_OutboundSemiProduct,
-            FlexiStockClient.MaterialWarehouseId => WarehouseDocumentType_OutboundMaterial,
-            _ => throw new InvalidOperationException("Unknown warehouse for consumption movement for warehouseId " + warehouseId)
-        };
-        return documentType;
-    }
-
-    private static string CreateDescription(SubmitManufactureClientRequest request)
-    {
-        return request.ManufactureType == ErpManufactureType.Product && request.Items.Count == 1
-            ? $"{request.Items[0].ProductCode} - {request.Items[0].ProductName}"
-            : request.ManufactureInternalNumber;
     }
